requirements.txt
PyYAML==3.10
externals
virtualenv
temp_dir
zope.dottedname==4.0.0
